+++ docs/make.jl
@@ -13,7 +13,7 @@ makedocs(
     authors = "Éric Thiébaut and contributors",
     modules = [QuickHeaps],
     pages = ["index.md", "install.md", "binaryheaps.md", "priorityqueues.md",
-             "nodes.md", "customize.md", "library.md"],
+             "customize.md", "library.md"],
     checkdocs = :exports,
 )